fix: cleanly split reducer and context
